Refactor /Auth
