--- frontend/src/features/Auth/SignIn/index.ts
@@ -1,1 +1,0 @@
-export { SignIn } from "./SignIn"
